+++ common/task/src/cancellation/manager.rs
@@ -1,7 +1,9 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::event::SentStatus;
+use crate::cancellation::tracker::{Cancelled, ShutdownTracker};
+use crate::spawn::JoinHandle;
+use crate::ShutdownToken;
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
 use log::error;
@@ -9,19 +11,7 @@ use std::future::Future;
 use std::mem;
 use std::pin::Pin;
 use std::time::Duration;
-use thiserror::Error;
-use tokio_util::sync::{
-    CancellationToken, DropGuard, WaitForCancellationFuture, WaitForCancellationFutureOwned,
-};
-use tokio_util::task::TaskTracker;
-use tracing::{debug, info, trace, warn};
-
-use crate::spawn::{spawn_named_future, JoinHandle};
-use crate::spawn_future;
-use tokio::task::JoinSet;
-
-#[cfg(unix)]
-use tokio::signal::unix::{signal, SignalKind};
+use tracing::info;
 
 #[cfg(not(target_arch = "wasm32"))]
 use tokio::time::sleep;
@@ -29,95 +19,19 @@ use tokio::time::sleep;
 #[cfg(target_arch = "wasm32")]
 use wasmtimer::tokio::sleep;
 
-pub const DEFAULT_MAX_SHUTDOWN_DURATION: Duration = Duration::from_secs(5);
+#[cfg(unix)]
+use tokio::signal::unix::{signal, SignalKind};
+use tokio::task::JoinSet;
 
-#[derive(Debug, Error)]
-#[error("task got cancelled")]
-pub struct Cancelled;
+#[allow(deprecated)]
+pub struct ShutdownManager {
+    pub(crate) legacy_task_manager: Option<crate::TaskManager>,
 
-#[derive(Debug, Clone, Default)]
-pub struct ShutdownToken {
-    inner: CancellationToken,
-}
+    pub(crate) shutdown_signals: ShutdownSignals,
 
-impl ShutdownToken {
-    /// Leave the drop in no-op replacement for `send_status_msg` for easier migration from `TaskClient`.
-    #[deprecated]
-    #[track_caller]
-    pub fn send_status_msg(&self, status: SentStatus) {
-        let caller = std::panic::Location::caller();
-        warn!("{caller} attempted to send {status} - there are no more listeners of those");
-    }
+    pub(crate) tracker: ShutdownTracker,
 
-    pub fn new() -> Self {
-        ShutdownToken {
-            inner: CancellationToken::new(),
-        }
-    }
-
-    pub fn ephemeral() -> Self {
-        ShutdownToken::default()
-    }
-
-    pub fn inner(&self) -> &CancellationToken {
-        &self.inner
-    }
-
-    pub fn child_token(&self) -> ShutdownToken {
-        ShutdownToken {
-            inner: self.inner.child_token(),
-        }
-    }
-
-    pub fn cancel(&self) {
-        self.inner.cancel();
-    }
-
-    /// Returns `true` if the `ShutdownToken` is cancelled.
-    pub fn is_cancelled(&self) -> bool {
-        self.inner.is_cancelled()
-    }
-
-    pub fn cancelled(&self) -> WaitForCancellationFuture<'_> {
-        self.inner.cancelled()
-    }
-
-    pub fn cancelled_owned(self) -> WaitForCancellationFutureOwned {
-        self.inner.cancelled_owned()
-    }
-
-    // Returned guard will cancel this token (and all its children) on drop unless disarmed.
-    pub fn drop_guard(self) -> ShutdownDropGuard {
-        ShutdownDropGuard {
-            inner: self.inner.drop_guard(),
-        }
-    }
-
-    pub async fn run_until_cancelled<F>(&self, fut: F) -> Option<F::Output>
-    where
-        F: Future,
-    {
-        self.inner.run_until_cancelled(fut).await
-    }
-
-    pub async fn run_until_cancelled_owned<F>(self, fut: F) -> Option<F::Output>
-    where
-        F: Future,
-    {
-        self.inner.run_until_cancelled_owned(fut).await
-    }
-}
-
-pub struct ShutdownDropGuard {
-    inner: DropGuard,
-}
-
-impl ShutdownDropGuard {
-    pub fn disarm(self) -> ShutdownToken {
-        ShutdownToken {
-            inner: self.inner.disarm(),
-        }
-    }
+    pub(crate) max_shutdown_duration: Duration,
 }
 
 #[derive(Default)]
@@ -127,224 +41,6 @@ impl ShutdownSignals {
     pub async fn wait_for_signal(&mut self) {
         self.0.join_next().await;
     }
-}
-
-/// Extracted [`TaskTracker`] and [`ShutdownToken`] to more easily allow tracking nested tasks
-/// without having to pass whole [`ShutdownManager`] around
-#[derive(Clone, Default, Debug)]
-pub struct ShutdownTracker {
-    root_cancellation_token: ShutdownToken,
-
-    // the reason I'm not using a `JoinSet` is because it forces us to use futures with the same `::Output` type
-    tracker: TaskTracker,
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-impl ShutdownTracker {
-    #[track_caller]
-    pub fn spawn<F>(&self, task: F) -> JoinHandle<F::Output>
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        let tracked = self.tracker.track_future(task);
-        spawn_future(tracked)
-    }
-
-    #[track_caller]
-    pub fn try_spawn_named<F>(&self, task: F, name: &str) -> JoinHandle<F::Output>
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        trace!("attempting to spawn task {name}");
-        let tracked = self.tracker.track_future(task);
-        spawn_named_future(tracked, name)
-    }
-
-    #[track_caller]
-    pub fn spawn_on<F>(&self, task: F, handle: &tokio::runtime::Handle) -> JoinHandle<F::Output>
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        self.tracker.spawn_on(task, handle)
-    }
-
-    #[track_caller]
-    pub fn spawn_local<F>(&self, task: F) -> JoinHandle<F::Output>
-    where
-        F: Future + 'static,
-        F::Output: 'static,
-    {
-        self.tracker.spawn_local(task)
-    }
-
-    #[track_caller]
-    pub fn spawn_blocking<F, T>(&self, task: F) -> JoinHandle<T>
-    where
-        F: FnOnce() -> T,
-        F: Send + 'static,
-        T: Send + 'static,
-    {
-        self.tracker.spawn_blocking(task)
-    }
-
-    #[track_caller]
-    pub fn spawn_blocking_on<F, T>(&self, task: F, handle: &tokio::runtime::Handle) -> JoinHandle<T>
-    where
-        F: FnOnce() -> T,
-        F: Send + 'static,
-        T: Send + 'static,
-    {
-        self.tracker.spawn_blocking_on(task, handle)
-    }
-
-    /// Spawn the task that will get cancelled if a global shutdown signal is detected
-    #[track_caller]
-    pub fn try_spawn_named_with_shutdown<F>(
-        &self,
-        task: F,
-        name: &str,
-    ) -> JoinHandle<Result<F::Output, Cancelled>>
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        trace!("attempting to spawn task {name} (with top-level cancellation)");
-
-        let caller = std::panic::Location::caller();
-        let shutdown_token = self.clone_shutdown_token();
-        let name_owned = name.to_string();
-        let tracked = self.tracker.track_future(async move {
-            match shutdown_token.run_until_cancelled_owned(task).await {
-                Some(result) => {
-                    debug!("{name_owned} @ {caller}: task has finished execution");
-                    Ok(result)
-                }
-                None => {
-                    debug!("{name_owned} @ {caller}: shutdown signal received, shutting down");
-                    Err(Cancelled)
-                }
-            }
-        });
-        spawn_named_future(tracked, name)
-    }
-
-    /// Spawn the task that will get cancelled if a global shutdown signal is detected
-    #[track_caller]
-    pub fn spawn_with_shutdown<F>(&self, task: F) -> JoinHandle<Result<F::Output, Cancelled>>
-    where
-        F: Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        let caller = std::panic::Location::caller();
-        let shutdown_token = self.clone_shutdown_token();
-        self.tracker.spawn(async move {
-            match shutdown_token.run_until_cancelled_owned(task).await {
-                Some(result) => {
-                    debug!("{caller}: task has finished execution");
-                    Ok(result)
-                }
-                None => {
-                    trace!("{caller}: shutdown signal received, shutting down");
-                    Err(Cancelled)
-                }
-            }
-        })
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-impl ShutdownTracker {
-    #[track_caller]
-    pub fn spawn<F>(&self, task: F) -> JoinHandle<F::Output>
-    where
-        F: Future + 'static,
-    {
-        let tracked = self.tracker.track_future(task);
-        spawn_future(tracked)
-    }
-
-    #[track_caller]
-    pub fn try_spawn_named<F>(&self, task: F, name: &str) -> JoinHandle<F::Output>
-    where
-        F: Future + 'static,
-    {
-        let tracked = self.tracker.track_future(task);
-        spawn_named_future(tracked, name)
-    }
-
-    /// Spawn the task that will get cancelled if a global shutdown signal is detected
-    #[track_caller]
-    pub fn try_spawn_named_with_shutdown<F>(
-        &self,
-        task: F,
-        name: &str,
-    ) -> JoinHandle<Result<F::Output, Cancelled>>
-    where
-        F: Future<Output = ()> + Send + 'static,
-    {
-        let caller = std::panic::Location::caller();
-        let shutdown_token = self.clone_shutdown_token();
-        let tracked = self.tracker.track_future(async move {
-            match shutdown_token.run_until_cancelled_owned(task).await {
-                Some(result) => {
-                    debug!("{caller}: task has finished execution");
-                    Ok(result)
-                }
-                None => {
-                    trace!("{caller}: shutdown signal received, shutting down");
-                    Err(Cancelled)
-                }
-            }
-        });
-        spawn_named_future(tracked, name)
-    }
-
-    /// Spawn the task that will get cancelled if a global shutdown signal is detected
-    #[track_caller]
-    pub fn spawn_with_shutdown<F>(&self, task: F) -> JoinHandle<Result<F::Output, Cancelled>>
-    where
-        F: Future<Output = ()> + Send + 'static,
-    {
-        let caller = std::panic::Location::caller();
-        let shutdown_token = self.clone_shutdown_token();
-        let tracked = self.tracker.track_future(async move {
-            match shutdown_token.run_until_cancelled_owned(task).await {
-                Some(result) => {
-                    debug!("{caller}: task has finished execution");
-                    Ok(result)
-                }
-                None => {
-                    trace!("{caller}: shutdown signal received, shutting down");
-                    Err(Cancelled)
-                }
-            }
-        });
-        spawn_future(tracked)
-    }
-}
-
-impl ShutdownTracker {
-    pub fn child_shutdown_token(&self) -> ShutdownToken {
-        self.root_cancellation_token.child_token()
-    }
-
-    pub fn clone_shutdown_token(&self) -> ShutdownToken {
-        self.root_cancellation_token.clone()
-    }
-}
-
-#[allow(deprecated)]
-pub struct ShutdownManager {
-    legacy_task_manager: Option<crate::TaskManager>,
-
-    shutdown_signals: ShutdownSignals,
-
-    tracker: ShutdownTracker,
-
-    max_shutdown_duration: Duration,
 }
 
 // note: default implementation will ONLY listen for SIGINT and will ignore SIGTERM and SIGQUIT
